Add openvswitch binaries to nova
Nova calls the openvswitch binary when neutron is using the openvswitch
driver to attach the port to the vswitch. This is not a known config
ahead of time and is dependant on a change in the neutron services.

This is an exception to the rule that can be removed if and when nova
switches to python ovs

Change-Id: I30d8251eec2e21b8181f8902d7899f8339ced6a1